+++ src/main/java/hu/elte/inf/szofttech2023/team3/spacewar/model/building/SolarPowerPlant.java
@@ -10,7 +10,7 @@ import hu.elte.inf.szofttech2023.team3.spacewar.model.space.objects.Planet;
  */
 public class SolarPowerPlant extends Building {
     // TODO: modify the efficiency consistently
-    private double efficiency = planet.temperature;
+    private int efficiency = planet.temperature;
 
     public SolarPowerPlant(Planet planet) {
         super(planet);
@@ -22,12 +22,12 @@ public class SolarPowerPlant extends Building {
         efficiency = efficiencyOfLevel();
     }
 
-    public double exportEnergy() {
+    public int exportEnergy() {
         return efficiency;
     }
 
     // TODO: modify the efficiency of a level
-    public double efficiencyOfLevel() {
+    public int efficiencyOfLevel() {
         return level + planet.temperature ;
     }
